+++ pylum/grating_coupler/gc_clean.py
@@ -1,6 +1,7 @@
-import lumapi
+if __name__ == "__main__":
+    import lumapi
 
-s = lumapi.FDTD()
-s.load("grating_coupler_2D.fsp")
-s.switchtolayout()
-s.save()
+    s = lumapi.FDTD()
+    s.load("grating_coupler_2D.fsp")
+    s.switchtolayout()
+    s.save()
