+++ pylum/grating_coupler/test_sweep.yml
@@ -5,15 +5,14 @@ GC_draw.lsf: "# 2D Grating Coupler Model\n\n# Draw GC\nredrawoff;\ngap=period*(1
   \ GC teeth;\nfor(i=0:gc_number)\n{\n  addrect;\n  set('name','GC_tooth');\n  set('material',material);\n\
   \  set('y',0.5*thick_Si);\n  set('y span',thick_Si);\n  set('x min',gap+i*period);\n\
   \  set('x max',period+i*period);\n}\nselectpartial('GC');\naddtogroup('GC');\n\n\
-  \n# draw silicon substrate;\naddrect;\nset('name','Si_sub');\nset('material','Si\
-  \ (Silicon) - Dispersive & Lossless');\nset('x max',30e-6);\nset('x min', -20e-6);\n\
-  set('y',-1*(thick_BOX+0.5*Si_substrate));\nset('y span',Si_substrate);\nset('alpha',0.2);\n\
-  \n#draw burried oxide;\naddrect;\nset('name','BOX');\nset('material','SiO2 (Glass)\
-  \ - Const');\nset('x max',30e-6);\nset('x min',-20e-6);\nset('y min',-thick_BOX);\n\
-  set('y max',thick_Clad);\nset('override mesh order from material database',true);\n\
-  set('mesh order',3);\nset('alpha',0.3);\n\n#draw waveguide;\naddrect;\nset('name','WG');\n\
-  set('material','Si (Silicon) - Dispersive & Lossless');\nset('x min',-20e-6);\n\
-  set('x max', 0);\nset('y',0.11e-6);\nset('y span',0.22e-6);\n"
+  \n# draw silicon substrate;\naddrect;\nset('name','Si_sub');\nset('material', material);\n\
+  set('x max',30e-6);\nset('x min', -20e-6);\nset('y',-1*(thick_BOX+0.5*Si_substrate));\n\
+  set('y span',Si_substrate);\nset('alpha',0.2);\n\n#draw burried oxide;\naddrect;\n\
+  set('name','BOX');\nset('material', material_clad);\nset('x max',30e-6);\nset('x\
+  \ min',-20e-6);\nset('y min',-thick_BOX);\nset('y max',thick_Clad);\nset('override\
+  \ mesh order from material database',true);\nset('mesh order',3);\nset('alpha',0.3);\n\
+  \n#draw waveguide;\naddrect;\nset('name','WG');\nset('material', material);\nset('x\
+  \ min',-20e-6);\nset('x max', 0);\nset('y',0.11e-6);\nset('y span',0.22e-6);\n"
 GC_init.lsf: '# define grating coupler parameters
 
   period = 6.6e-07;
@@ -35,9 +34,9 @@ GC_init.lsf: '# define grating coupler parameters
 
   Si_substrate = 2e-06;
 
-  materials;
+  material = ''Si (Silicon) - Palik'';
 
-  material = "Si (Silicon) - Dispersive & Lossless";
+  material_clad = ''SiO2 (Glass) - Palik'';
 
   width_wg = 5e-07;
 
@@ -113,72 +112,15 @@ GC_sweeps.lsf: "# Sweep various parameters of the grating coupler\n\nnewproject;
   }\n\nWL=linspace(lambda-0.5*wl_span,lambda+0.5*wl_span,frequency_points);\n\nfor(jj=1:5)\n\
   {\n    plot(WL, abs(M_T(1:frequency_points,jj)));\n    holdon;\n}\n?10*log10(max(abs(M_T)));\
   \  # lowest insertion loss\nholdoff;\nplot(M_sweep, abs(M_Tlambda));\n\nmatlabsave('results');"
+function_name: sweep
 main.lsf: 'GC_init;
 
   GC_sweeps;'
-materials.lsf: '# materials.lsf - creates a dispersive material model in Lumerical.
-
-  matname = "Si (Silicon) - Dispersive & Lossless";
-
-  newmaterial = addmaterial("Lorentz");
-
-  setmaterial(newmaterial,"name",matname);
-
-  setmaterial(matname,"Permittivity",7.98737492);
-
-  setmaterial(matname,"Lorentz Linewidth",1e8);
-
-  setmaterial(matname,"Lorentz Resonance",3.93282466e+15);
-
-  setmaterial(matname,"Lorentz Permittivity",3.68799143);
-
-  setmaterial(matname,"color",[0.85, 0, 0, 1]); # red
-
-
-  matname = "Air (1)";
-
-  if (1) { newmaterial = addmaterial("Dielectric"); }
-
-  setmaterial(newmaterial,"name",matname);
-
-  setmaterial(matname,"Refractive Index",1);
-
-  setmaterial(matname,"color",[0.85, 0.85, 0, 1]);
-
-
-  matname = "SiO2 (Glass) - Dispersive & Lossless";
-
-  newmaterial = addmaterial("Lorentz");
-
-  setmaterial(newmaterial,"name",matname);
-
-  setmaterial(matname,"Permittivity",2.119881);
-
-  setmaterial(matname,"Lorentz Linewidth",1e10);
-
-  setmaterial(matname,"Lorentz Resonance",3.309238e+13);
-
-  setmaterial(matname,"Lorentz Permittivity", 49.43721);
-
-  setmaterial(matname,"color",[0.5, 0.5, 0.5, 1]); # grey
-
-
-  matname = "SiO2 (Glass) - Const";
-
-  newmaterial = addmaterial("Dielectric");
-
-  setmaterial(newmaterial,"name",matname);
-
-  setmaterial(matname,"Permittivity",1.444^2);
-
-  setmaterial(matname,"color",[0.5, 0.5, 0.5, 1]); # grey
-
-  '
 name: sweep
-function_name: sweep
 settings.json: '{"period": 6.6e-07, "ff": 0.5, "n_gratings": 50, "wg_height": 2.2e-07,
   "etch_depth": 7e-08, "box_height": 2e-06, "clad_height": 2e-06, "substrate_height":
-  2e-06, "material": "Si (Silicon) - Dispersive & Lossless", "wg_width": 5e-07, "polarization":
-  "TE", "wavelength": 1.55e-06, "gc_position": 4.5e-06, "fiber_angle_deg": 20, "draw_source_script_name":
-  "GC_setup_Gaussian", "sweep_variable": "period", "sweep_start": 6.2e-07, "sweep_stop":
-  7e-07, "sweep_points": 5}'
+  2e-06, "material": "Si (Silicon) - Palik", "material_clad": "SiO2 (Glass) - Palik",
+  "wg_width": 5e-07, "polarization": "TE", "wavelength": 1.55e-06, "gc_position":
+  4.5e-06, "fiber_angle_deg": 20, "draw_source_script_name": "GC_setup_Gaussian",
+  "sweep_variable": "period", "sweep_start": 6.2e-07, "sweep_stop": 7e-07, "sweep_points":
+  5}'
